Refactor powdertracker to be in mining
